(Refact) seperate download button
