+++ src/entities/resume/download/ui/Form.tsx
@@ -1,21 +1,19 @@
 'use client';
 
-import { Button } from '@/src/widgets/ui/button';
 import { StateBasedModal } from '@/src/widgets/ui/modal';
 import GridFormat from '@/src/entities/resume/download/ui/GridFormat';
+import { DownloadButton } from '@/src/features/resume/download/ui/downloadButton';
 
 import { findTier } from '@/src/entities/resume/download/lib/findTier';
 import { useResumeResultStore } from '@/src/features/resume/form/model/resumeStore';
 import { handleDownload } from '@/src/features/resume/download/model/handleDownload';
 
 import React from 'react';
-import { useRouter } from 'next/navigation';
 
 export default function Form() {
-  const router = useRouter();
   const { result: data } = useResumeResultStore();
 
-  const { isSuccess, setIsSuccess, isFailed, setIsFailed, requestResume } = handleDownload();
+  const { isSuccess, setIsSuccess, isFailed, setIsFailed } = handleDownload();
   return (
     <>
       <div className="gap-0">
@@ -57,31 +55,8 @@ export default function Form() {
         <hr />
       </div>
 
-      <div className="gap-3 flex flex-col">
-        <div className="flex w-full gap-auto justify-between mt-4 px-4">
-          <Button
-            label="이력서 초안 생성"
-            variant="secondary"
-            className="w-[11.5625rem] rounded-lg text-xs font-medium"
-            onClick={requestResume}
-          ></Button>
-          <Button
-            label="고급 이력서 생성(Coming Soon)"
-            variant="primary"
-            disabled={true}
-            className="w-[11.5625rem] rounded-lg text-xs font-medium"
-          ></Button>
-        </div>
-        <Button
-          label="홈으로"
-          variant="link"
-          fullWidth={true}
-          className="text-xs font-medium underline"
-          onClick={() => {
-            router.push('/');
-          }}
-        ></Button>
-      </div>
+      <DownloadButton />
+
       <StateBasedModal
         open={isSuccess}
         onOpenChange={() => {}}
